requirements.txt
autocorrect==2.6.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
editdistpy==0.1.3
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.21.3
idna==3.6
ijson==3.2.3
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
psycopg2==2.9.9
pyspellchecker==0.8.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sympy==1.12
symspellpy==6.7.7
textblob==0.18.0.post0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
